Ensure a basically functioning standard browser if proclient off r=mossop
As part of this work I wanted to bring browser.xhtml back to how it was prior
to the MR2 work. As a sanity check for that I wanted to make sure the JS wasn't
breaking or throwing errors when I switched to use the normal browser.xhtml. It
seemed like we were going to have to do this eventually, so may as well get
started on it now.

Differential Revision: https://phabricator.services.mozilla.com/D115846