Revert browser.xhtml changes and create browser-proclient.xhtml r=mossop
This patch basically exists to make the next one easier to read. There is a lot
of duplication between browser.xhtml and browser-proclient.xhtml in this patch,
which the following patch aims to consolidate.

Regarding the introduction of the pref and env var, I'm not tied to this, I
just thought that the most flexible possible thing we could do was to make this
a runtime configurable change. If we end up wanting this to be a completely
separate binary and everything, it's easier to switch to that from a runtime
variable than the other way around.

Depends on D115846

Differential Revision: https://phabricator.services.mozilla.com/D115847