Add browser-proclient.xhtml and ifdefed changes to included files r=mossop
There were a few options here, as discussed in the issue. I went down the route
of two xhtml files which include shared elements, together with preprocessor
directives to handle small differences, because 1) it is the most flexible, and
2) it should have no performance cost.

I additionally took the opportunity to go through already modified include
files and wrap their changes in preprocessor directives. This does not account
for css changes and JS changes, somewhat obviously, but is intended to get us
partway back to being able to merge this stuff back with m-c.

Depends on D116503

Differential Revision: https://phabricator.services.mozilla.com/D115848