bug 1380123 - mitigate duplicate nightly graphs off the same revision
We can still get multiple nightly graphs off the same revision, if the 2nd cron task runs before the 1st one finishes. However, once the 1st finishes, at least in theory we will prevent a 2nd nightly graph. (This has the side effect of not being able to trigger a 2nd nightly graph off the same revision, even if the first nightly graph failed, but I think that's still an improvement. We can work around this with a new push, or triggering the platform-specific nightly graphs.)