MR2-242 - Allow resizing and persisting size of companion r=jaws
This patch will allow resizing of the companion sidebar, and will persist that
size using the xulstore. Note that this will not resize the top area of the
core UI which seems intended to match the size of this. Given the mockups, it
looks like we'll want to have that area be a part of the sidebar rather than a
part of the top bar, so I'd like to defer doing something fancy to resize that
until we are sure what the DOM structure of all of this should be.

Differential Revision: https://phabricator.services.mozilla.com/D119331