MR2-197 - Collapse the River favicons when they're overflowing. r=mossop
The mechanism for overlapping the View elements in the River intentionally
uses hardcoded CSS values, since much of the Active View Manager is likely
going to be refactored and rearchitected in an upcoming series of patches.

This is just an intermediary step to demonstrate the overlapping / spread
behaviour that's described by the UX specifications.

Depends on D119678

Differential Revision: https://phabricator.services.mozilla.com/D119679