MR2-572 - Make sure the TopLevelNavigationDelegate ignores the special OAuth2 tab. r=mkaply
The TopLevelNavigationDelegate works to try to make most top-level
navigations result in a new <browser> being created. OAuth2 creates
a special browser that it expects to eventually redirect to a localhost
URI with a token from the authentication provider. The
TopLevelNavigationDelegate was confusing it because that redirect was
resulting in a new <browser> that it had no knowledge of.

This patch makes it so that the OAuth2 <browser> is specially configured
to never create new <browser>'s for top-level navigations. This configuration
uses sharedData in order to persist across cross-domain process flips.

Differential Revision: https://phabricator.services.mozilla.com/D123141