MR2-589 - Don't delegate location.replace loads in TopLevelNavigationDelegateChild. r=mossop
The location.replace() API replaces the current history entry with one
that points at the URI passed to location.replace(). This prevents going
back to a replaced location.

The TopLevelNavigationDelegateChild was not detecting this kind of load,
and was causing a new <browser> to be created for them. This resulted in
more Views being shown in the River than necessary, since the replaced
Views should probably be subsumed by the things replacing them.

Differential Revision: https://phabricator.services.mozilla.com/D123368