MR2-613 - Properly update the SharedData Set for the TopLevelNavigationDelegate if a Set already exists. r=mkaply
The SharedData structure that we use to propagate data down to content
processes seems to only have an update triggered when the value that
a SharedData key points at has its reference updated. Before this
patch, if we'd already used OAuth2.jsm to authenticate, then on a
subsequent authentication attempt, we were getting at a pre-existing
Set and adding an element to it. That's not enough for SharedData to
notice. Instead, we write a new Set each time we add a new browserId
to it.

Differential Revision: https://phabricator.services.mozilla.com/D123361