MR2-628 - Only register TopLevelNavigationDelegate once at startup. r=jhirsch
As part of its constructor, GlobalHistory attempts to register the
TopLevelNavigationDelegate JSWindowActor if
browser.tabs.openNewTabForMostNavigations is set to true. Unfortunately,
this means that when another window is opened, the new GlobalHistory
instance also tries to register that same actor, which results in an
exception being thrown in the construction of GlobalHistory.

This patch adds a guard to prevent re-registering the JSWindowActor.
This is a stopgap until we enable
browser.tabs.openNewTabForMostNavigations by default, since then we
can find a more suitable place to do the registration that doesn't
cause a pref read (probably at the top of the GlobalHistory module).

Differential Revision: https://phabricator.services.mozilla.com/D123513