MR2-746 - Add the animation for setting aside a session r=jhirsch
This adds the animation for setting aside a session. This does not implement
any of the logic of actually setting aside the session and getting a new one.
Additionally, this animation drops frames. The problem here is in the way
WebRender handles scale animations, in that we have to re-rasterize glyphs
for every frame of the scale transition. We will have to add a way to suppress
this - that, however, will be in a future patch.

Differential Revision: https://phabricator.services.mozilla.com/D125122