MR2-587 - Don't delegate top-level navigations when loads are triggered by the System Principal. r=mossop,bdahl
In order to keep our tests working as expected with the TopLevelNavigationDelegate
enabled, we're going to let system-triggered navigations to proceed within the
same <browser> rather than tryint to make new one be created. This is so that
tests of the form:

```
BrowserTestUtils.loadURI(browser, "http://example.com");
await BrowserTestUtils.browserLoaded(browser);
```

will continue to work.

This also adjusts a pre-existing pinebuild test so that it uses the System Principal
to trigger a navigation.

Differential Revision: https://phabricator.services.mozilla.com/D125279