MR2-747 - Correct multideck RemoveFrame logic r=mconley
It seems that somewhere along the way, this hack stopped being necessary to get
view transition animations to work. I no longer see the previous frame getting
a transient RemoveFrame on view transition. So I'm going to just go ahead and
remove the hack, but I am going to make sure we adjust mPrevIndex if the
removed frame is below it, for the same reason that we do that with mIndex.

Differential Revision: https://phabricator.services.mozilla.com/D125737