MR2-793 - Export interactions to pyodide r=mossop
The exported interactions.json is transformed to a pyodide-friendly format and a small python script is run on the data.
A pandas dataframe is created from the column names and the row values.
The shape and column names are logged from python and returned to the runtime.

This patch also fixes missing resources from vendor/pyodide.

Differential Revision: https://phabricator.services.mozilla.com/D125894