MR2-950 - Avoid firing SSWindowStateReady when dealing with cached SHEntries in GlobalHistory. r=mossop
We were using SessionStore.setTabState to deal with the case when a cached SHEntry needed
to be used to navigate to a View.

SessionStore.setTabState ends up using the SessionStore.restoreTab mechanism, which in turn
causes the SSWindowStateBusy event to fire on the window, which makes GlobalHistory think
we're doing a full session restoration, which seems to greatly confuse things when we're
dealing with the cached SHEntry.

To work around this, instead of using SessionStore.restoreTab, this patch makes GlobalHistory
use SessionHistory.restoreFromParent to revive the cached SHEntry. This avoids the event,
while still letting us deserialize the cached SHEntry and inserting it into the sessionHistory
of the <browser>.

Differential Revision: https://phabricator.services.mozilla.com/D126721