MR2-958 - Exclude active sessions from SessionManager.query results. r=mak
Most of the time we won't want a list to have an active session displayed as restoring the session or looking at your active session is not much use. Therefore we should skip active sessions from the results by default - but keep an option for the interactions viewer and if we find a case elsewhere.

Differential Revision: https://phabricator.services.mozilla.com/D126701