MR2-960 - Make it easier to rearrange pinned Views. r=jhirsch
Before this patch, PinnedViews assumed that the dragged ViewGroup contained
the current View, which is not necessarily the case. This fixes that case.

It also makes it so that dropping results in pinning the selected View _after_
the pinned View that the dragged View is dragged over. This is probably what
we should have been doing in the first place, especially considering that
dragging over uses CSS to create a space _after_ the dragged-over ViewGroup,
and not _before_ it.

Differential Revision: https://phabricator.services.mozilla.com/D127911