MR2-885 - Start Pro Client at MacOS system startup r=mr2-core-ui-reviewers,mconley
This adds a launchd agent for the proclient based on the pref
browser.startup.launchOnOSLogin. I used a launchd agent instead of a login
item because we don't get a permissions alert when we try to add a launchd
agent for the user, and because we need to provide environment variables
anyway in order to get it to run in the background. One of these environment
variables (MOZ_APP_NO_DOCK) won't matter until the next merge from m-c, by
the way.

Differential Revision: https://phabricator.services.mozilla.com/D128820