MR2-1346 - Fix refresh() promise and add added event test coverage. r=calixte
The problem with the refresh() promise was that during the (offline, fast)
"plan" phase of task execution, complex tasks like the
`defineAtMostOnceTask`-defined `sync_refresh` can decide to use a different
task id for their (online, async/slow) execution phase.  Specifically, they
change to use the name of the task group that they'll use.  And even if
they didn't do this, their "at most once" semantics implement coalescing so
that a series of "sync_refresh" calls that haven't executed yet can all
accumulate into a single request.  In that case, we inherently must end up
using a different name.

I apparently had thought of some of these ramifications, as the planning phase
is allowed to return a value, and we specifically return the/a task group
promise from the planning phase.  This will work for both coalesced and
uncoalesced tasks.

And so for sync_refresh, we wait for that.

I've updated tests to use this as appropriate, specifically
test_calendar_search_all.js which was instead waiting for multiple "seeked"
events.

test_calendar_single_day had originally intended to add events after
performing the first sync, but it ran into the inverted date range helper
bug and then potentially the database bug.  With Calixte having fixed the
range inversion bug, we can now safely add events, and this is cleanly done
thanks to the refresh() method existing above.  (Which is also why these two
somewhat orthogonal changes are in the same patch; they both help test each
other.)

Differential Revision: https://phabricator.services.mozilla.com/D130835