MR2-1350 - Do more to prevent GlobalHistory from accessing gBrowser until its ready. r=bgrins
Before this patch, because active-view-manager.js is a module script, it was possible for
the ActiveViewManager custom element to have its script parsed and connectedCallback fire
before DOMContentLoaded had fired. Part of the connectedCallback was accessing the
gGlobalHistory lazy getter, which caused GlobalHistory to construct. That constructor
would then attempt to attach listeners to things like the window's gBrowser.

The problem is that sometimes at that point, gBrowser isn't yet available. This would
cause GlobalHistory construction to fail and a bunch of things to break.

This patch does two things:

1. Separates GlobalHistory construction from initialization. This is similar to how
   gBrowser works. This allows GlobalHistory to be constructed and accessed regardless
   if gBrowser is around.
2. Moves the initialization routines for GlobalHistory into a method that's called
   by browser.js after we _know_ that gBrowser is available.
3. As a belt-and-suspenders measure, makes ActiveViewManager inherit from MozHTMLElement
   and use the delayConnectedCallback mechanism to defer running the connectedCallback
   until DOMContentLoaded has fired. This helps ensure that the ActiveViewManager code
   can access all of the DOM nodes that it needs inside of its connectedCallback.

Differential Revision: https://phabricator.services.mozilla.com/D130933