MR2-1316 - Take the companion browser out of the browsers messagemanagergroup. r=mstriemer
The messagemanagergroup attribute allows <browser> elements to identify that they're
part of a particular messaging group. This allows certain groups of <browser>'s to
get JSWindowActors registered for them (via the messageManagerGroups property during
registration).

The "browsers" group is a poorly named group from the e10s-days, but it's for <browser>
elements within tabbrowser.

The Companion <browser> is a sidebar, and should not be in that group. I discovered this
when I realized that the PageWireframeChild actor was getting registered for it.

Differential Revision: https://phabricator.services.mozilla.com/D131384