MR2-1315 - Add back button gestures for entering the history carousel. r=mstriemer,cmkm
I'm bolting on support for this kind of long-press onto the pre-existing
gClickAndHoldListenersOnElement object in browser.js, though that originally was
only for opening context menus automatically.

I removed the "backForwardMenu" context attributes from the back/forward buttons
for pinebuild because:

a) They were only showing up when the currently staged <browser> actually had
   internal back/forward state, which is something we're trying to disguise a
   bit with the AVM / River

b) They interfered with the long-press gesture for entering the history carousel.

Differential Revision: https://phabricator.services.mozilla.com/D131387