MR2-1370 - Switch Pro Client to Inter (part 1) r=mconley
So, this is the coarse pass of switching to Inter. There's plenty of other
remaining pieces, but this does the bulk of the work. I am putting this in
its own patch now since I think this poses the hairiest question remaining in
this project: do we want to apply this via * selectors? I think that this is
certainly an ugly solution, but I also think it's the best option. The two
alternatives I can think of are 1) overriding the system-ui generic at the
platform level to use Inter when possible, or 2) using many targeted CSS
selectors all over the frontend.

Regarding 1), I don't think this is actually crazy, but given the complexity
of implementing the system-ui generic in the first place, I think this is
likely to actually be a bit complicated. We would also not want to affect
content, which means we'd need to only override the system-ui generic in
privileged contexts. I think that's likely doable, but I just didn't go down
that route as the * selector didn't seem to pose real practical problems. I'm
very open to debate on this.

Regarding 2), I think it's just asking for a bunch of weird problems and
annoyances for anyone adding new UI, where they will have to remember to change
the font to Inter. I don't think this is sustainable and I can't think of any
clear advantage to doing this.

Let me know if that all sounds reasonable.

The remaining work here is to adjust the layouts of things that have become off
as a result of this switch, and then to catch any areas which are not covered
by this change, if we want to do so. Poking around, some pieces like devtools
are still not covered by this change. I am not convinced that we *want* them to
be covered by this change, however. I will ask in the Jira ticket.

Differential Revision: https://phabricator.services.mozilla.com/D131428