MR2-1426 - Move activationPolicy-related incantations to nsCocoaWindow::Show r=mstange
So, it turns out that we get some weird intermittent focus behavior when we put this
inside applicationShouldHandleReopen. Moving this to sandwich the showing of the
window seems to do the trick. It looks like what was happening is we were activating
our application before we actually had a visible window, which was resulting in
our window sometimes being in the front but our application not active. I'm not
100% sure on a story for what was going here, but this change does appear to fix
the issue.

Differential Revision: https://phabricator.services.mozilla.com/D132057