MR2-1497 - Make getViewGroups and getPinnedViewGroups responsible for waiting for LitElement updates to complete. r=prathiksha
I suspect that not waiting for LitElement to finish updating the DOM is causing the
frequent failures that this bug is associated with. This patch makes the PinebuildTestUtils
helper functions for getting at the ViewGroups responsible for waiting for the DOM
to finish updating, which should hopefully make it easier to write more resilient
tests.

Differential Revision: https://phabricator.services.mozilla.com/D132970