MR2-1387 - Disable back/forward while entering history carousel, but allow them to work once the transition is complete. r=mossop,prathiksha
I debated with myself a bit about adding the #currentHistoryCarouselInternalView, or just re-using the
#currentInternalView member. I ended up going with the former route because of how we would need to
call `setView` once exiting the carousel, and if #currentInternalView already matched the selected
view, we would bail out early.

I could have added a "force" option to `setView`, or somehow made it apparent to `setView` that we
needed to actually set the view as current, but I opted to add the new member and make the back/forward
stuff a little more complicated in order to not make `setView` any more complicated than it already
is.

Differential Revision: https://phabricator.services.mozilla.com/D133654