MR2-1598 - Don't show pinned views in the history carousel. r=prathiksha
This hides pinned views from showing up in the history carousel. We do this
instead of just filtering out pinned views because the indices of the preview
array are (currently) set up to always match the indices of the GlobalHistory
#viewStack, and hiding pinned previews was simpler than trying to translate
indexes with an offset at the parent/child communication layer.

This also updates some tests to include the pinned view case.

Differential Revision: https://phabricator.services.mozilla.com/D134067