MR2-1534 - Allow transparent content in privileged about pages r=emilio
Part of the upcoming Project Lavender UI redesign is a shared background for
the whole browser in certain scenarios. For example, after setting aside your
session, the entire background of the browser animates to a purpley image
which is shared by the new session set aside page as well as by the companion
(certion UI elements in the companion sit on top of this purpley background.)
We could implement this as an image in each of the display lists, just offset
so there is no visible seam between them, or we could simply let certain
pages have transparent content, letting the global browser background show
through. This patch implements the latter solution.

Differential Revision: https://phabricator.services.mozilla.com/D133826