MR2-1662 - Do a better job of not showing pinned views in the history carousel. r=cmkm
The original ticket for this (MR2-1598) was resolved with a patch that caused us to
hide the pinned views in the history carousel. This was done to keep the patch simple,
but subsequent testing revealed problems with the original approach.

The first problem was that we were setting visibility: hidden on the pinned previews,
rather than display: none. This meant that they still took up space and could still be
scrolled to with the trackpad.

The second problem was that we have some CSS for the history carousel which assumes
that the first preview element is visible and takes up space. This allows us to show
the edges of the previous and next previews when we've got a preview selected that
isn't the first or last.

I decided to bite the bullet here and go for the more robust but slightly more
complicated patch. What we're doing here is decoupling the indexes of the array
for previews that the history carousel gets with the indexes of the array of the
views that GlobalHistory is managing. This means explicitly including the View
index in each preview that comes down to the history carousel, and using that value
for selection / identification rather than the index of the previews array (or the
index of the location within the preview element list).

This is probably the right thing to do anyways - as GlobalHistory gets more complex,
it probably makes more sense to have history carousel not know so much on the shape
of the whole View array.

Differential Revision: https://phabricator.services.mozilla.com/D136184