MR2-1422 - Return to login list if login modified behind-the-scenes r=jhirsch
When a user logs into a website with a stored login item,
the password manager updates that login item automatically
in order to increment the usage counter for that item.

This triggers a modified login event. In m-c, this has the
result of taking the user out the login edit view, if they
happen to be in the edit view at the time, and returning to
the read-only view of that login item. This is fine in m-c.

However, in pine, we don't have a read-only view, and the
only way to see the URL associated with a login item is to
be in that item's edit view. So, when the modified login
event is intercepted, and the component attempts to exit the
edit view, this was having the effect of breaking the login
view entirely, since there's no read-only view to return to.

In order to avoid this issue, in pine, we'll return the user
to the login list view when a modified login event occurs.

This is identical to the behavior that's currently taken by a
user-initiated modified login event: the user makes a change,
clicks the save button, and the user returns to the login list.

If we had access to the full details of the login item which has
been modified, we might be able to detect whether the only
change that's been made is the change to increment the usage
counter, and keep the edit view open in that case. However, we
don't have access to this level of detail, so we can't make that
comparison. This solution is a compromise to avoid breakage.

Differential Revision: https://phabricator.services.mozilla.com/D136409