MR2-1955 - Make companion calendar tests more reliable. r=hjones
We have a number of frequent failures in automation right now related to
calendar tests. Digging into it, it looks like some of our calendar tests
were setting state on MozLitElement's, and then not long after, querying
the DOM to test that it was in the expected state.

The problem here is that Lit batches state changes and doesn't immediately
update the DOM. This means that _sometimes_, the DOM wouldn't be updated
by the time the querying occurred, and the test would fail. It would really
depend on the refresh driver for painting, since that's the trigger for
Lit to actually update the DOM.

This patch makes it so that we wait for the `updateComplete` Promise on
the calendar list MozLitElement after setting calendar events. That Promise
is resolved after the DOM has been updated.

Differential Revision: https://phabricator.services.mozilla.com/D139723