MR2-1955 - Move the mouse cursor to the toolbar in between companion browser tests. r=hjones
This fixes a frequent failure in browser_calendar_join_meeting.js where the Join Meeting
button on an event was expected to be hidden. After investigating, it seems what happened
is that a previous test in the folder (browser_PageActionMenu_copy.js) synthesized the mouse
at the center of the Page Action Menu's Copy button, which (I believe) placed it right at
the corner of the companion window such that, somehow, the :hover state was set on the
Join Meeting button causing it to appear.

I think this started to rear its head only because of recent style changes that probably
positioned the Page Action Menu and Companion items just so.

I cannot be 100% certain why the :hover pseudoclass applies even though the place where
the event's button would normally be is not directly under the mouse, but moving the
mouse back over thei toolbar after the browser_PageActionMenu_copy.js test seems
to satisfy the test. I decided to just do this between every test to try to avoid
hitting this in the future.

Depends on D139723

Differential Revision: https://phabricator.services.mozilla.com/D139756