MR2-476 - Disable the prefs and tabs sync engines in ProClient. r=jhirsch
For ProClient, we don't want to sync any prefs, so we completely disable the pref sync engine. We also don't want tab syncing, so we disable that as well.
To prevent "Send to tab" being used to send tabs to ProClient, we set the prefs "identity.fxaccounts.commands.enabled" to false.

This patch also moves the flow-reset homepage pref back inside the main
PINEBUILD-specific ifdef at the bottom of the prefs file, setting the
default back to about:blank for future m-c mergeability.

Differential Revision: https://phabricator.services.mozilla.com/D141694