MR2-1686 - Correctly display the urlbar in fullscreen view r=prathiksha
Some of the hacks I needed to get the urlbar moving with the rest of the toolbar seem to no longer work, and removing them fixes the issue. I imagine it's an upstream change, but haven't dug in to find the specific commit. Fingers crossed this resolves the issue permanently.

Compare to https://phabricator.services.mozilla.com/D141295. All we need to do now is set the "shifted" attribute so that the background of the toolbar contents isn't transparent when it shifts downward.

Differential Revision: https://phabricator.services.mozilla.com/D142623