MR2-1768 - Manually create a ContentPrincipal for an InternalView when setting up a lazy-loaded browser after session restoration. r=prathiksha
This patch updates InternalView to always use the URI of the associated View when assigning
the content principal. This also fixes an issue where the content principal could be wrong
if the underlying browser isn't actually pointed at the URI at the time of restoration (which
can occur for lazily-loaded Views).

Differential Revision: https://phabricator.services.mozilla.com/D143386