MR2-2263 - Remove old or redundant tests now that TopLevelNavigationDelegate is being removed. r=jhirsch
browser_StageManager_TopLevelNavigationDelegate.js directly tested the
TopLevelNavigationDelegate behaviour, and will not be needed now that we're
removing TopLevelNavigationDelegate.

browser_StageManager_targetTopLevelLinkClicksToBlank.js tests the new mechanism,
but a version of this test landed in mozilla-central as part of the platform
support, and currently lives at
docshell/test/browser/browser_targetTopLevelLinkClicksToBlank.js.

Depends on D143729

Differential Revision: https://phabricator.services.mozilla.com/D143730