MR2-2374 - Use classes instead of IDs for Workspace DOM items. r=jhirsch
When Workspace support is enabled, we can create multiple <workspace> elements,
which clones the items from the workspace-template as direct children. Those items
currently have IDs, which means if there are multiple workspaces then we have DOM
items rooted at the same root document node with the same IDs, which we shouldn't
do.

This patch changes the workspace template (and the various workspace scripts and
styles) to use classes instead.

Differential Revision: https://phabricator.services.mozilla.com/D144611