MR2-2165 - Follow-up: Invert HCM colors for hovered toolbarbuttons. r=cmkm
This makes sure that in HCM, we have ButtonFace as the toolbarbutton
background, and ButtonText for stroke and fill of toolbarbutton borders
and icons respectively. Upon :hover, those colours are inverted. On :active,
SelectedItem is applied as the background colour, with ButtonText again being
used for stroke and fill.

This patch does not include changes to the Page Action Menu button, which
will be addressed in a follow-up.

Differential Revision: https://phabricator.services.mozilla.com/D144484