MR2-2000 - Remove missing workspaceId workaround. r=mossop
This workaround was added in https://hg.mozilla.org/projects/pine/rev/8877f3a07e72a23a43059c46f37e303ade7835fc
for MR2-1959 because we needed to be able to handle older session stored
on the disk that didn't have workspaceIds set.

Sessions are designed to expire after a month, and we've been stashing
workspaceIds in sessions for over a month now (since MR2-1715 introduced
it). This means that there shouldn't be any users with sessions that are
this old anymore, meaning we should be able to remove the workaround.

It's likely that, over time, we're going to change the schema of the
session state object. To avoid having to do this kind of month-long
workaround, we should add some kind of schema versioning system. I
will be doing that in a later patch in this stack for MR2-2223.

Differential Revision: https://phabricator.services.mozilla.com/D144958